Delete button to general components
