+++ src/configuration/Buttons.js
@@ -1,8 +1,6 @@
 import IconButton from "@material-ui/core/IconButton";
 import UndoIcon from "@material-ui/icons/Undo";
-import Tooltip from "@material-ui/core/Tooltip";
 import Button from "@material-ui/core/Button";
-import DeleteIcon from "@material-ui/icons/Delete";
 import React from "react";
 import { makeStyles } from "@material-ui/core/styles";
 import NodesClient from "../api/NodesClient";
@@ -77,27 +75,4 @@ function UpdateConfigurationButton({
   );
 }
 
-function DeleteButton({ size, tooltip, classes, onClick, disabled }) {
-  const btn = () => (
-    <IconButton
-      color="secondary"
-      aria-label="delete"
-      size={size || "small"}
-      className={classes}
-      onClick={onClick}
-      disabled={disabled}
-    >
-      <DeleteIcon />
-    </IconButton>
-  );
-  if (tooltip) {
-    return (
-      <Tooltip title={tooltip}>
-        <div>{btn()}</div>
-      </Tooltip>
-    );
-  }
-  return btn();
-}
-
-export { RestoreConfigurationButton, UpdateConfigurationButton, DeleteButton };
+export { RestoreConfigurationButton, UpdateConfigurationButton };
